// src/components/lib/types.ts
// main axis
export type Align =
  | "start"
  | "center"
  | "end"
  | "space-around"
  | "space-between"
  | "space-evenly"
// cross axis
export type CrossAlign = "start" | "center" | "end" | "baseline" | "stretch"
// gap
export type Gap = "sm" | "md" | "lg" | "xl"
// tone color
export type Tone = "default" | "subdued" | "success" | "critical"
// text type
export type Variant =
  | "headingXs"
  | "headingSm"
  | "headingMd"
  | "headingLg"
  | "headingXl"
  | "heading2xl"
  | "heading3xl"
  | "bodyXs"
  | "bodySm"
  | "bodyMd"
  | "bodyLg"
// text font weight
export type FontWeight = "regular" | "medium" | "semibold" | "bold"

// page width
export type PageWidth =
  | "defaultWidth"
  | "narrowWidth"
  | "formWidth"
  | "fullWidth"